Remove hearing selection page.
